requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database & Auth
supabase==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI & LLM (Conflict-free versions)
ibm-watsonx-ai==1.0.0
openai==1.10.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
httpx==0.26.0
python-multipart==0.0.6
websockets==12.0

# Optional for frontend
react==18.2.0
zustand==4.4.7
tailwindcss==3.3.0
@supabase/supabase-js==2.38.4